+++ demo/routes.nim
@@ -1,0 +1,8 @@
+import kraut/sugar
+
+
+routes:
+  "#/": index
+  "#/users": users
+  "#/users/{userId}": user
+